requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
configobj==5.0.6
Flask>=1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
importlib-metadata==1.7.0
itsdangerous==0.24
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-gnupg==0.4.6
python-ldap==3.3.0
secrets==1.0.2
six==1.15.0
SQLAlchemy==1.3.17
systematic==4.8.7
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
